count unwrapped scripts as running
